Restructured the directories to avoid merging conflicts
