+++ Backend-Sys/.env
@@ -1,0 +1,9 @@
+# Configuring POSTGRESQL DB
+DBUSER=data_shield
+DBNAME=data_shield_db
+DBPASS=data_shield_pass
+DBHOST=localhost
+DBPORT=5432
+
+# SECURITY WARNING: keep the secret key used in production secret!
+SECRET_KEY='django-insecure-^+w4a4-rikfv@ph3k4e#5juiul_3^7)4ie_ac6=3o5u1dwfzse'
