--- Susan_G/Backend-sys/.gitignore
@@ -1,1 +1,0 @@
-env/
