--- Susan_G/Backend-sys/backend_dev/backend_dev/urls.py
@@ -1,8 +1,0 @@
-
-from django.contrib import admin
-from django.urls import path, include
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('',include('test_backend_app.urls')),
-]
