--- Susan_G/Backend-sys/backend_dev/templates/home.html
@@ -1,1 +1,0 @@
-<h1>Mum </h1>
