--- Susan_G/Backend-sys/backend_dev/test_backend_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TestBackendAppConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'test_backend_app'
