--- Susan_G/Backend-sys/backend_dev/test_backend_app/templates/test_backend_app/home.html
@@ -1,1 +1,0 @@
-<h1>Hi </h1>
