--- Susan_G/Backend-sys/backend_dev/test_backend_app/templates/test_backend_app/login_register.html
@@ -1,34 +1,0 @@
-<!-- login_register.html -->
-
-<div class="container">
-    <h2>Login or Register</h2>
-
-    <!-- User Login Form -->
-    <div class="login-section">
-        <h3>Login</h3>
-        <form method="post">
-            {% csrf_token %}
-            <div>
-                <label for="email">Email:</label>
-                <input type="email" name="email" required>
-            </div>
-            <div>
-                <label for="password">Password:</label>
-                <input type="password" name="password" required>
-            </div>
-            <button type="submit">Login</button>
-        </form>
-        <p>Don't have an account? <a href="#register">Register here</a></p>
-    </div>
-
-    <!-- User Registration Form -->
-    <div class="registration-section" id="register">
-        <h3>Register</h3>
-        <form method="post">
-            {% csrf_token %}
-            {{ form.as_p }}
-            <button type="submit">Register</button>
-        </form>
-        <p>Already have an account? <a href="#login">Login here</a></p>
-    </div>
-</div>
