--- Susan_G/Backend-sys/backend_dev/test_backend_app/templates/test_backend_app/organization_registration.html
@@ -1,5 +1,0 @@
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <button type="submit">Create Organization</button>
-</form>
