--- Susan_G/Backend-sys/requirements.txt
@@ -1,1 +1,0 @@
-Python 3.8.2
